+++ R/workflow.R
@@ -24,6 +24,7 @@ gcap.workflow <- function(tumourseqfile, normalseqfile,
                           genome_build = c("hg38", "hg19"),
                           model = "XGB32",
                           tightness = 1L,
+                          gap_cn = 4L,
                           outdir = getwd(),
                           result_file_prefix = paste0("gcap_", uuid::UUIDgenerate(TRUE)),
                           allelecounter_exe = "~/miniconda3/envs/cancerit/bin/alleleCounter",
@@ -183,7 +184,7 @@ gcap.workflow <- function(tumourseqfile, normalseqfile,
   lg$info("====================================")
   lg$info("Step 4: Run scoring and summarizing")
   lg$info("====================================")
-  out <- gcap.runScoring(model_input, genome_build, tightness = tightness)
+  out <- gcap.runScoring(model_input, genome_build, tightness = tightness, gap_cn = gap_cn)
 
   save_file <- file.path(outdir, paste0(result_file_prefix, "_prediction_result.rds"))
   lg$info("Saving raw prediction result to {save_file}")